+++ bindgen/src/program.ts
@@ -23,7 +23,7 @@ import path from "path";
 
 import { debug, enableDebugging } from "./debug";
 import { generate } from "./generator";
-import { OptInSpec, InvalidSpecError, parseOptInSpec, parseSpecs } from "./spec";
+import { OptInSpec, InvalidSpecError, parseOptInSpec, parseSpecs, generateSchema } from "./spec";
 import { Template, importTemplate } from "./templates";
 
 type GenerateOptions = {
@@ -109,12 +109,21 @@ program
         optInSpec = parseOptInSpec(optInPath);
       }
       generate({ rawSpec, optInSpec, template: await template, outputPath });
-      process.exit(0);
     } catch (err) {
       printError(err);
-      process.exit(1);
+      process.exitCode = 1;
     }
   });
+
+program.command("generate-schema").action(() => {
+  try {
+    console.log("Generating spec.schema.json");
+    generateSchema();
+  } catch (err) {
+    printError(err);
+    process.exitCode = 1;
+  }
+});
 
 program
   .command("validate")
@@ -129,9 +138,8 @@ program
     try {
       parseSpecs(specPaths);
       console.log(chalk.green("Validation passed!"));
-      process.exit(0);
     } catch (err) {
       printError(err);
-      process.exit(1);
+      process.exitCode = 1;
     }
   });